python_project/pyproject.toml
[tool.poetry]
name = "sesame-automate"
version = "0.1.0"
description = "Automation tool for Sesame Time tracking system"
authors = ["Marc Rabadan <marc.rabadan@telavox.com>"]
readme = "README.md"
packages = [{include = "sesame_automate"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
requests-mock = "^1.11.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sesame-automate = "sesame_automate.main:main"